Fix several issues
Bug #1: If a 17-symbol-long number was input at
the start, the calculator would not allow an
operator to be input by the user.

Bug #2: If the second term consisted of a minus
sign (-), pressing any of the operators would
make the calculator attempt to compute an
operation with "-" as the second term.